project-03-flashcards-ui/pyproject.toml
[project]
name = "project-03-flashcards-ui"
version = "0.1.0"
description = "AI Flashcard Generator + Study UI (Streamlit) with local memory + Anki export"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.36.0",
  "python-dotenv>=1.0.1",
  "langchain>=0.2.0",
  "langchain-openai>=0.1.8",
]

[tool.uv]
dev-dependencies = []
